Backend/requirements.txt
flask
flask-cors
requests
python-dotenv
google-generativeai>=0.8.0
gunicorn